# Speed up element arithmetic