--- a/news/arfcache.rst
+++ b/news/arfcache.rst
@@ -1,0 +1,3 @@
+**Fixed:**
+
+* cache arf conversions to speed up arithmetic (in particular with multiple generators)