by caching the RealNumber Arf conversion

```
                                                               Before     After
ElementBenchmark<IntegerRing>/addition_Z/0                        456 ns      456 ns
ElementBenchmark<IntegerRing>/addition_Z/1/1/1                    455 ns      455 ns
ElementBenchmark<IntegerRing>/addition_Z/2/1/0/0/1               2173 ns     2262 ns
ElementBenchmark<IntegerRing>/addition_Z/2/3/4/1/2               3882 ns     4001 ns
ElementBenchmark<RationalField>/addition_Q/0                      745 ns      749 ns
ElementBenchmark<RationalField>/addition_Q/1/1/1                  744 ns      749 ns
ElementBenchmark<RationalField>/addition_Q/2/1/0/0/1             3847 ns     4058 ns
ElementBenchmark<RationalField>/addition_Q/2/3/4/1/2             8704 ns     8899 ns
ElementBenchmark<NumberField>/addition_K/0                        318 ns      314 ns
ElementBenchmark<NumberField>/addition_K/1/1/1                    318 ns      314 ns
ElementBenchmark<NumberField>/addition_K/2/1/0/0/1               1862 ns     1656 ns
ElementBenchmark<NumberField>/addition_K/2/3/4/1/2               3031 ns     2523 ns
ElementBenchmark<IntegerRing>/multiplication_Z/0                 2245 ns     1799 ns
ElementBenchmark<IntegerRing>/multiplication_Z/1/1/1             2379 ns     1798 ns
ElementBenchmark<IntegerRing>/multiplication_Z/2/1/0/0/1         8357 ns     4098 ns
ElementBenchmark<IntegerRing>/multiplication_Z/2/3/4/1/2       308111 ns    23845 ns
ElementBenchmark<RationalField>/multiplication_Q/0               3646 ns     3051 ns
ElementBenchmark<RationalField>/multiplication_Q/1/1/1           3604 ns     3120 ns
ElementBenchmark<RationalField>/multiplication_Q/2/1/0/0/1      10942 ns     6817 ns
ElementBenchmark<RationalField>/multiplication_Q/2/3/4/1/2     330521 ns    36949 ns
ElementBenchmark<NumberField>/multiplication_K/0                 2290 ns     1837 ns
ElementBenchmark<NumberField>/multiplication_K/1/1/1             2239 ns     1834 ns
ElementBenchmark<NumberField>/multiplication_K/2/1/0/0/1         7748 ns     4111 ns
ElementBenchmark<NumberField>/multiplication_K/2/3/4/1/2       302580 ns    24447 ns
ElementBenchmark<IntegerRing>/truediv_Z/0                       11516 ns    10321 ns
ElementBenchmark<IntegerRing>/truediv_Z/1/1/1                   11599 ns    10407 ns
ElementBenchmark<IntegerRing>/truediv_Z/2/1/0/0/1               19527 ns    14830 ns
ElementBenchmark<IntegerRing>/truediv_Z/2/3/4/1/2             2923067 ns   176478 ns
ElementBenchmark<RationalField>/truediv_Q/0                     15024 ns    14329 ns
ElementBenchmark<RationalField>/truediv_Q/1/1/1                 15477 ns    14169 ns
ElementBenchmark<RationalField>/truediv_Q/2/1/0/0/1             27868 ns    22729 ns
ElementBenchmark<RationalField>/truediv_Q/2/3/4/1/2           3134787 ns   323036 ns
ElementBenchmark<NumberField>/truediv_K/0                       12185 ns    10763 ns
ElementBenchmark<NumberField>/truediv_K/1/1/1                   12175 ns    10839 ns
ElementBenchmark<NumberField>/truediv_K/2/1/0/0/1               20897 ns    14649 ns
ElementBenchmark<NumberField>/truediv_K/2/3/4/1/2             3039940 ns   170837 ns
ElementBenchmark<IntegerRing>/floordiv_Z/0                      13849 ns    12084 ns
ElementBenchmark<IntegerRing>/floordiv_Z/1/1/1                  13709 ns    12079 ns
ElementBenchmark<IntegerRing>/floordiv_Z/2/1/0/0/1              21988 ns    16664 ns
ElementBenchmark<IntegerRing>/floordiv_Z/2/3/4/1/2            3636616 ns   199936 ns
ElementBenchmark<RationalField>/floordiv_Q/0                    17103 ns    15819 ns
ElementBenchmark<RationalField>/floordiv_Q/1/1/1                17262 ns    15932 ns
ElementBenchmark<RationalField>/floordiv_Q/2/1/0/0/1            29566 ns    24474 ns
ElementBenchmark<RationalField>/floordiv_Q/2/3/4/1/2          3731512 ns   343992 ns
ElementBenchmark<NumberField>/floordiv_K/0                      14322 ns    12864 ns
ElementBenchmark<NumberField>/floordiv_K/1/1/1                  14325 ns    12926 ns
ElementBenchmark<NumberField>/floordiv_K/2/1/0/0/1              23033 ns    16809 ns
ElementBenchmark<NumberField>/floordiv_K/2/3/4/1/2            3630046 ns   184111 ns
ElementBenchmark<IntegerRing>/arb_Z/0                            2346 ns     1138 ns
ElementBenchmark<IntegerRing>/arb_Z/1/1/1                        2327 ns     1143 ns
ElementBenchmark<IntegerRing>/arb_Z/2/1/0/0/1                    2311 ns     1142 ns
ElementBenchmark<IntegerRing>/arb_Z/2/3/4/1/2                  289431 ns    13043 ns
ElementBenchmark<RationalField>/arb_Q/0                          2543 ns     1349 ns
ElementBenchmark<RationalField>/arb_Q/1/1/1                      2510 ns     1338 ns
ElementBenchmark<RationalField>/arb_Q/2/1/0/0/1                  2543 ns     1321 ns
ElementBenchmark<RationalField>/arb_Q/2/3/4/1/2                286573 ns    13563 ns
ElementBenchmark<NumberField>/arb_K/0                            2594 ns     1422 ns
ElementBenchmark<NumberField>/arb_K/1/1/1                        2583 ns     1424 ns
ElementBenchmark<NumberField>/arb_K/2/1/0/0/1                    2561 ns     1432 ns
ElementBenchmark<NumberField>/arb_K/2/3/4/1/2                  113436 ns     6106 ns
```